language fix

+++ app/Http/Controllers/Backend/Admin/TeachersController.php
@@ -238,7 +238,12 @@ class TeachersController extends Controller
     {
 
         $teacher = User::findOrFail($id);
-        $teacher->delete();
+
+        if ($teacher->courses->count() > 0) {
+            return redirect()->route('admin.teachers.index')->withFlashDanger(trans('alerts.backend.general.teacher_delete_warning'));
+        } else {
+            $teacher->delete();
+        }
 
         return redirect()->route('admin.teachers.index')->withFlashSuccess(trans('alerts.backend.general.deleted'));
     }
